Refactor: Add structured explanation for Claiming Candidate hint
This commit introduces a structured, multi-step explanation for the "Claiming Candidate" hint and refactors the underlying implementation for better clarity and maintainability.

**Key Changes:**

*   **`ClaimingCandidateExplanation.kt`**:
    *   The `ClaimingCandidateExplanation` class has been moved to its own dedicated file (`sudoku-core/src/main/java/io/github/mgrablo/sudokucore/hints/strategies/ClaimingCandidateExplanation.kt`).
    *   The logic for generating hint steps now uses `HintMessageFormatter` for cleaner string construction, replacing manual concatenation of `HintExplanationPart`s.

*   **String Resources**:
    *   New, more descriptive string keys (`hint_claiming_candidate_step_1` to `_4`) have been added to `feature/game/src/main/res/values/hint_strings.xml` to support the multi-step explanation.
    *   Corresponding `HintStringKey` enums have been added to `sudoku-core`.

*   **Hint System Integration**:
    *   The `AndroidHintStringProvider` has been updated to map the new `CLAIMING_CANDIDATE` string keys to their Android resource IDs.
    *   `ClaimingCandidateExplanation` is now correctly imported and integrated into the hint provider and factory modules.
