+++ sudoku-core/src/main/java/io/github/mgrablo/sudokucore/hints/HintExplanationFactory.kt
@@ -1,5 +1,6 @@
 package io.github.mgrablo.sudokucore.hints
 
+import io.github.mgrablo.sudokucore.hints.strategies.ClaimingCandidateExplanation
 import io.github.mgrablo.sudokucore.hints.strategies.HiddenSingleExplanation
 import io.github.mgrablo.sudokucore.hints.strategies.NakedSingleExplanation
