+++ sudoku-core/src/main/java/io/github/mgrablo/sudokucore/hints/HintExplanationStrategy.kt
@@ -13,116 +13,6 @@ interface HintExplanationStrategy {
 	): List<HintExplanationStep>
 }
 
-class ClaimingCandidateExplanation : HintExplanationStrategy {
-	override fun generateStructuredHintExplanation(
-		grid: SudokuGrid,
-		hint: Hint,
-		stringProvider: HintStringProvider,
-	): List<HintExplanationStep> {
-		val hintType = hint.type as HintType.ClaimingCandidate
-
-		// Calculate block id
-		val blockId =
-			(hint.row / grid.subgridSize) * grid.subgridSize + (hint.col / grid.subgridSize) + 1
-
-		// Organize enforcing cells by row or column
-		val groupType = hintType.groupType
-		val scopeType = if (groupType is GroupType.Row) ScopeType.ROW else ScopeType.COLUMN
-
-		// Get affected cells coordinates (1-indexed)
-		val affectedCells = hint.affectedCells
-			.sortedBy { it.row }
-			.map { Pair(it.row + 1, it.col + 1) }
-
-		// Group enforcing cells by row/column
-		val enforcingCellsByScope = hint.enforcingCells
-			.groupBy { if (groupType is GroupType.Row) it.row else it.col }
-			.map { (scopeIndex, cells) ->
-				scopeIndex to cells.map { Pair(it.row + 1, it.col + 1) }
-			}
-
-		// Build the explanation steps
-		val steps = mutableListOf<HintExplanationStep>()
-
-		// Step 1: Focus on block
-		steps.add(
-			HintExplanationStep(
-				listOf(
-					HintExplanationPart.Text(stringProvider.getString(HintStringKey.FOCUS_ON_BLOCK)),
-					HintExplanationPart.Text(" "),
-					HintExplanationPart.ScopeReference(ScopeType.BLOCK, blockId),
-					HintExplanationPart.Text("!"),
-				),
-			),
-		)
-
-		// Step 2: Explain for each row/column where the value can be placed
-		enforcingCellsByScope.forEach { (scopeIndex, cells) ->
-			steps.add(
-				HintExplanationStep(
-					listOf(
-						HintExplanationPart.Text(stringProvider.getString(HintStringKey.IN)),
-						HintExplanationPart.Text(" "),
-						HintExplanationPart.ScopeReference(scopeType, scopeIndex + 1),
-						HintExplanationPart.Text(", "),
-						HintExplanationPart.Value(hint.value),
-						HintExplanationPart.Text(" "),
-						HintExplanationPart.Text(stringProvider.getString(HintStringKey.CAN_ONLY_BE_PLACED_IN)),
-						HintExplanationPart.Text(" "),
-						HintExplanationPart.Text(stringProvider.getString(HintStringKey.FOLLOWING_CELLS)),
-						HintExplanationPart.Text(" "),
-						HintExplanationPart.CellCoordinatesGroup(cells),
-						HintExplanationPart.Text("."),
-					),
-				),
-			)
-		}
-
-		// Step 3: Conclusion - removing candidates from other cells in the block
-		steps.add(
-			HintExplanationStep(
-				listOf(
-					HintExplanationPart.Text(stringProvider.getString(HintStringKey.THUS_REMOVE_FROM)),
-					HintExplanationPart.Text(" "),
-					HintExplanationPart.Value(hint.value),
-					HintExplanationPart.Text(" "),
-					HintExplanationPart.Text(stringProvider.getString(HintStringKey.CANNOT_BE_PLACED_IN)),
-					HintExplanationPart.Text(" "),
-					HintExplanationPart.ScopeReference(ScopeType.BLOCK, blockId),
-					HintExplanationPart.Text("."),
-				),
-			),
-		)
-
-		// Step 4: Show affected cells
-		if (affectedCells.isNotEmpty()) {
-			steps.add(
-				HintExplanationStep(
-					listOf(
-						HintExplanationPart.Text(stringProvider.getString(HintStringKey.AFFECTED_CELLS_ARE)),
-						HintExplanationPart.Text(" "),
-						HintExplanationPart.CellCoordinatesGroup(affectedCells),
-						HintExplanationPart.Text("."),
-					),
-				),
-			)
-		}
-
-		// Step 5: Technique name
-		steps.add(
-			HintExplanationStep(
-				listOf(
-					HintExplanationPart.TechniqueName(
-						stringProvider.getString(HintStringKey.TECHNIQUE_CLAIMING_CANDIDATE),
-					),
-				),
-			),
-		)
-
-		return steps
-	}
-}
-
 class PointingCandidateExplanation : HintExplanationStrategy {
 	override fun generateStructuredHintExplanation(
 		grid: SudokuGrid,
