+++ sudoku-core/src/main/java/io/github/mgrablo/sudokucore/hints/HintStringProvider.kt
@@ -57,6 +57,12 @@ enum class HintStringKey {
 	HIDDEN_SINGLE_STEP_2,
 	HIDDEN_SINGLE_STEP_3,
 
+	// Claiming Candidate
+	CLAIMING_CANDIDATE_STEP_1,
+	CLAIMING_CANDIDATE_STEP_2,
+	CLAIMING_CANDIDATE_STEP_3,
+	CLAIMING_CANDIDATE_STEP_4,
+
 	// Scope parts
 	ROW,
 	COLUMN,
